# api/hqbase-mail-api-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HQBase Mail API",
    "version": "2.0.0",
    "description": "Stable v2 mail API for HQBase clients, automations, and agents. Additive fields may be introduced within v2; clients must ignore unknown response fields. Administrative APIs are not part of this contract."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://mail.example.com",
          "description": "HQBase origin"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v2/changes": {
      "get": {
        "summary": "Read message changes",
        "description": "Returns a bounded page of message upserts and deletion tombstones in journal order. A request without a cursor returns an empty checkpoint at the current high-water sequence.",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "A bounded page of message changes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageChangePage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_CHANGE_CURSOR` when the cursor is malformed, foreign, or beyond the current journal. `INVALID_CHANGE_FILTER` when a mailbox, folder, or search filter is supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "`CHANGE_CURSOR_EXPIRED` when future journal retention removes a required range. The client must start a new full bootstrap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Changes"],
        "operationId": "listMessageChanges",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from a checkpoint or previous changes page. Omit it only to create a new checkpoint."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of journal entries read for this page."
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
...
    "/api/v2/drafts": {
      "get": {
        "summary": "List drafts",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "One page of drafts, newest update first.",
            "headers": {
              "Link": {
                "required": false,
                "description": "RFC 8288 link to the next page, as `<url>; rel=\"next\"`. The header is absent on the last page.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Draft"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_DRAFT_CURSOR` when the cursor is malformed or is not a draft-list cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "listDrafts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of draft rows read for this page. Live access filtering can make the returned array shorter."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page. Read it from the Link header and send it unchanged."
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:send"]
      },
      "post": {
        "summary": "Create a draft",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "createDraft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              },
              "example": {
                "mailboxId": null,
                "replyToMessageId": null,
                "forwardOfMessageId": null,
                "from": "",
                "to": [],
                "cc": [],
                "bcc": [],
                "subject": "",
                "text": "",
                "html": ""
              }
            }
          }
        },
        "x-hqbase-agent-capabilities": ["mail:send"]
      }
    },
    "/api/v2/drafts/changes": {
      "get": {
        "summary": "Read draft changes",
        "description": "Returns a bounded page of private draft upserts and deletion tombstones in journal order. A request without a cursor returns an empty checkpoint at the current user-specific high-water sequence.",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "A bounded page of draft changes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftChangePage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_DRAFT_CHANGE_CURSOR` when the cursor is malformed, foreign, or beyond the user's current journal. `INVALID_DRAFT_CHANGE_FILTER` when a mailbox, folder, search, or timestamp filter is supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "`DRAFT_CHANGE_CURSOR_EXPIRED` when future journal retention removes a required range. The client must start a new full draft bootstrap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "listDraftChanges",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from a draft checkpoint or previous draft-changes page. Omit it only to create a new checkpoint."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of draft journal entries read for this page."
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:send"]
      }
    },
    "/api/v2/drafts/{id}": {
      "get": {
        "summary": "Get a draft",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "getDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:send"]
      },
      "patch": {
        "summary": "Update a draft",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Draft version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "updateDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              },
              "example": {
                "mailboxId": null,
                "replyToMessageId": null,
                "forwardOfMessageId": null,
                "from": "",
                "to": [],
                "cc": [],
                "bcc": [],
                "subject": "Updated draft",
                "text": "",
                "html": "",
                "version": 1
              }
            }
          }
        },
        "x-hqbase-agent-capabilities": ["mail:send"]
      },
      "delete": {
        "summary": "Delete a draft",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Draft deleted"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "deleteDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:send"]
      }
    },
    "/api/v2/drafts/{id}/attachments": {
      "post": {
        "summary": "Add a draft attachment",
        "description": "Records the file part's Content-Type, or application/octet-stream when no type is present. One file and all files in one draft can each total at most 25 MiB; larger uploads return 413 ATTACHMENTS_TOO_LARGE.",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "201": {
            "description": "Added attachment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftAttachment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "File or draft attachment total exceeds 25 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "addDraftAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "contentEncoding": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "contentType": "*/*"
                }
              }
            }
          }
        },
        "x-hqbase-agent-capabilities": ["mail:send"]
      }
    },
    "/api/v2/drafts/{draftId}/attachments/{id}": {
      "delete": {
        "summary": "Remove a draft attachment",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Attachment removed"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "removeDraftAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:send"]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Audience-bound OAuth token. Request resource={origin}/api/v2.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "refreshUrl": "/api/auth/oauth2/token",
            "scopes": {
              "mail:read": "Read accessible mail",
              "mail:write": "Organize accessible mail",
              "mail:send": "Manage drafts and send mail",
              "offline_access": "Request a refresh token"
            }
          }
        }
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Same-origin HQBase web session. Secure deployments may prefix the cookie name."
      },
      "agentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hqb_agent_<secret>",
        "description": "Mailbox agent credential created in Settings > Agents. It works only for the assigned Mail API capabilities and mailbox grant. Each operation lists required capabilities in x-hqbase-agent-capabilities."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MessageChange": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/MessageChangeUpsert"
          },
          {
            "$ref": "#/components/schemas/MessageChangeDelete"
          }
        ]
      },
      "MessageChangePage": {
        "type": "object",
        "required": ["changes", "nextCursor", "hasMore"],
        "properties": {
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageChange"
            }
          },
          "nextCursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}